requirements.txt
# Web Scraping
selenium==4.16.0
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# WebDriver Management
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Vector Database
qdrant-client==1.7.0

# LLM & Embeddings
google-genai==1.61.0
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.36.0

# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.0
pydantic-settings==2.1.0
python-multipart==0.0.9

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Logging
colorlog==6.8.0
